requirements.txt
astroid==2.2.5
autopep8==1.4.4
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-PyMongo==2.3.0
Flask-WTF==0.14.2
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
pymongo==3.9.0
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.5
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
